+++ src/components/SharedComponents/helpers/downloadAndShareHelpers/createDownloadLink.tsx
