+++ src/components/TeacherComponents/UnitDownloadButton/UnitDownloadButton.tsx
@@ -12,7 +12,7 @@ import { z } from "zod";
 import useUnitDownloadExistenceCheck from "../hooks/downloadAndShareHooks/useUnitDownloadExistenceCheck";
 
 import createAndClickHiddenDownloadLink from "@/components/SharedComponents/helpers/downloadAndShareHelpers/createAndClickHiddenDownloadLink";
-import { createUnitDownloadLink } from "@/components/SharedComponents/helpers/downloadAndShareHelpers/createDownloadResourcesLink";
+import { createUnitDownloadLink } from "@/components/SharedComponents/helpers/downloadAndShareHelpers/createDownloadLink";
 
 // teacher-unit-downloads experiment A/B test group keys and test values
 const variantKey = z
